Connected! (kind of)
Incorporated Anna's and Angel's new code.  It now works to a point: the
most recent rel will be added to the map, but only that rel.  When you
add a rel, you'll see that the info is added to the map as expected for
a moment, then the map reloads and the previously added rel disappears.
If you add only one concept (concept 1 or 2, doesn't matter), the map
will continue adding those, linking them to a blank node, without
reloading.